+++ src/shared/theme/components/BodyWrapper.js
@@ -7,42 +7,26 @@ const defaultStyle = {
   minWidth:  '100%',
   minHeight: '100%',
   wordWrap: 'break-word',
-  background: 'yellow',
 };
 
-const wrapperSpacing = {
-  paddingTop: 16,
-  paddingBottom: 16,
-};
-
-const textSpacing = {
-  paddingTop: 1,
-  paddingBottom: 1,
-};
-
-const BodyWrapper = ({ header, message, style }) => (
+const BodyWrapper = ({ child, style }) => (
   <div
     style={{
       ...defaultStyle,
       ...style
     }}
   >
-    <div style={wrapperSpacing}>
-      {header && <div style={textSpacing}>{header}</div>}
-      {message && <div style={textSpacing}>{message}</div>}
-    </div>
+    {child}
   </div>
 );
 
 BodyWrapper.propTypes = {
-  header: PropTypes.element,
-  message: PropTypes.element,
+  child: PropTypes.element,
   style: PropTypes.object,
 };
 
 BodyWrapper.defaultProps = {
-  header: null,
-  message: null,
+  child: null,
   style: {},
 };